+++ NewZhongYan/NewZhongYan/LocalMetaDataManager.m
@@ -66,8 +66,8 @@
             NSString* upt = metaData.updateTime ? metaData.updateTime :[NSDate curerntTime];
             NSInteger lcv = [metaData version];
             sql = [NSString stringWithFormat:
-                   @"INSERT OR REPLACE INTO DATA_VER (UPT,LCV,SID,OWUID) VALUES ('%@',%d,'%@','%@');",
-                   upt,lcv,metaData.dataCode,uid];
+                   @"INSERT OR REPLACE INTO DATA_VER (UPT,LCV,SID,OWUID) VALUES ('%@',%ld,'%@','%@');",
+                   upt,(long)lcv,metaData.dataCode,uid];
             
         }
         else if ([metaData isInitDataSnapped])
@@ -101,16 +101,16 @@
 {
     NSString* sql = [NSString stringWithFormat:
                      @"select AID from T_DOCUMENTS where strftime('%%s','now','start of day','-8 hour','-1 day') <= strftime('%%s',crtm) AND CHANNELID in (%@) and ENABLED = 1 AND READED = 0;",currentFid];
-    int count = [[DBQueue sharedbQueue] CountOfQueryWithSQL:sql];
-    return count ? [NSString stringWithFormat:@"%d",count] : nil;
+    NSInteger count = [[DBQueue sharedbQueue] CountOfQueryWithSQL:sql];
+    return count ? [NSString stringWithFormat:@"%ld",(long)count] : nil;
 }
 
 +(NSString*)newECMMeettingItemCount:(NSString*)currentFid
 {
     NSString* sql = [NSString stringWithFormat:
                      @"select  AID from T_DOCUMENTS WHERE DATETIME(EDTM) > DATETIME('now','localtime') and CHANNELID in (%@) and ENABLED = 1;",currentFid];
-    int count = [[DBQueue sharedbQueue] CountOfQueryWithSQL:sql];
-    return count ? [NSString stringWithFormat:@"%d",count] : nil;
+    NSInteger count = [[DBQueue sharedbQueue] CountOfQueryWithSQL:sql];
+    return count ? [NSString stringWithFormat:@"%ld",(long)count] : nil;
 }
 
 +(NSString*)newDataItemCount:(LocalDataMeta*)metaData;
@@ -145,7 +145,7 @@
                AND READED = 0;",
                [metaData localName],[datacode substringFromIndex:7]];
     }
-    int count = [[DBQueue sharedbQueue] CountOfQueryWithSQL:sql];
-    return count ? [NSString stringWithFormat:@"%d",count] : nil;
+    NSInteger count = [[DBQueue sharedbQueue] CountOfQueryWithSQL:sql];
+    return count ? [NSString stringWithFormat:@"%ld",(long)count] : nil;
 }
 @end
